Support nowiki tags, closes #14
This is a bit of a hack, as the <nowiki> tag is parsed out entirely by mwlib so we can't ever see it was there (we just get plaintext paragraphs containing unescaped wiki content, which makes sense for most uses.

So we wrap any <nowiki> tags in a magic marker HTML tag that gets passed through as plaintext. Kind of crummy, but the <mediawiki> tag is a macro-style tag even to medawiki so it shouldn't break under any valid Mediawiki syntax content.